feat(homepage): watermark stacks behind devices, softer glow halo
Watermark stacking:
- Previously the homepage watermark was rendered fixed at z-15 (above
  <main>), so its 6% blue tint painted on top of every device, making
  phone/browser screens look slightly transparent (visible JotilLabs
  hexagon "behind" the chat bubbles).
- Lifted the watermark out of BrandBackground into a new HomeWatermark
  client component, rendered as the first child of <main> on the
  homepage only. It now sits at z-1 inside <main>'s z-10 stacking
  context (= z-11 absolute), so:
    - section backgrounds (z-auto) sit BELOW the watermark -> watermark
      visible across the whole homepage as intended.
    - device-frame elements get z-index: 2 in globals.css -> their
      opaque bg-white screens cover the watermark, no more bleed-through.

Glow halo tone-down:
- DeviceGlow base opacities: outer 0.55 -> 0.32, inner 0.35 -> 0.18.
- Pulse keyframes: outer 0.45-0.75 -> 0.25-0.42, inner 0.25-0.5 -> 0.13-0.26.
- Slightly softer gradient color stops + larger blur radii.
- Net result: ~40-45% subtler halo. Mockup itself stays the focus, the
  multi-color wash reads as a soft accent rather than competing.
