+++ components/sections/showcase/devices/DeviceGlow.jsx
@@ -10,9 +10,9 @@ export function DeviceGlow({ radius = 60, inset = -18, intensity = 1 }) {
           inset,
           borderRadius: radius,
           background:
-            'linear-gradient(135deg, rgba(6,182,212,0.55), rgba(59,130,246,0.55), rgba(139,92,246,0.6), rgba(217,70,239,0.55))',
-          filter: 'blur(36px)',
-          opacity: 0.55 * intensity,
+            'linear-gradient(135deg, rgba(6,182,212,0.5), rgba(59,130,246,0.5), rgba(139,92,246,0.55), rgba(217,70,239,0.5))',
+          filter: 'blur(40px)',
+          opacity: 0.32 * intensity,
         }}
       />
       {/* Inner halo — tighter, sharper colour */}
@@ -24,9 +24,9 @@ export function DeviceGlow({ radius = 60, inset = -18, intensity = 1 }) {
           inset: typeof inset === 'number' ? inset + 8 : inset,
           borderRadius: radius,
           background:
-            'linear-gradient(135deg, rgba(6,182,212,0.7), rgba(139,92,246,0.7), rgba(217,70,239,0.6))',
-          filter: 'blur(14px)',
-          opacity: 0.35 * intensity,
+            'linear-gradient(135deg, rgba(6,182,212,0.6), rgba(139,92,246,0.6), rgba(217,70,239,0.5))',
+          filter: 'blur(16px)',
+          opacity: 0.18 * intensity,
         }}
       />
     </>
